+++ framework/base/ActionFilter.php
@@ -28,6 +28,8 @@ class ActionFilter extends Behavior
      * Note that if the filter is attached to a module, the action IDs should also include child module IDs (if any)
      * and controller IDs.
      *
+     * @since 2.0.9 action IDs can be specified as wildcards, e.g. `site/*`.
+     *
      * @see except
      */
     public $only;
@@ -138,6 +140,27 @@ class ActionFilter extends Behavior
     protected function isActive($action)
     {
         $id = $this->getActionId($action);
-        return !in_array($id, $this->except, true) && (empty($this->only) || in_array($id, $this->only, true));
+
+        if (empty($this->only)) {
+            $onlyMatch = true;
+        } else {
+            $onlyMatch = false;
+            foreach ($this->only as $pattern) {
+                if (fnmatch($pattern, $id)) {
+                    $onlyMatch = true;
+                    break;
+                }
+            }
+        }
+
+        $exceptMatch = false;
+        foreach ($this->except as $pattern) {
+            if (fnmatch($pattern, $id)) {
+                $exceptMatch = true;
+                break;
+            }
+        }
+
+        return !$exceptMatch && $onlyMatch;
     }
 }